break out muni filter and clean up
